refactor: moved events from tags module to recipes module
